Django4.2/requirements.txt
django==4.2.18
djangorestframework==3.15.2
jamo==0.4.1